+++ src/Stryker.CLI/Stryker.CLI/FileConfigReader.cs
@@ -61,6 +61,7 @@ namespace Stryker.CLI
             inputs.IgnoredMethodsInput.SuppliedInput = config.IgnoreMethods;
 
             inputs.ReportFileNameInput.SuppliedInput = config.ReportFileName;
+            inputs.BreakOnInitialTestFailureInput.SuppliedInput = config.BreakOnInitialTestFailure;
         }
 
         private static FileBasedInput LoadConfig(string configFilePath)