+++ src/Stryker.Core/Stryker.Core/Options/Inputs/SolutionInput.cs
@@ -1,6 +1,7 @@
-using Stryker.Core.Exceptions;
-using System.IO;
 using System.IO.Abstractions;
+using System.Linq;
+using System.Text;
+using Stryker.Core.Exceptions;
 
 namespace Stryker.Core.Options.Inputs
 {
@@ -10,11 +11,11 @@ namespace Stryker.Core.Options.Inputs
 
         protected override string Description => "Full path to your solution file. Required on dotnet framework.";
 
-        public string Validate(IFileSystem fileSystem)
+        public string Validate(string basePath, IFileSystem fileSystem)
         {
             if (SuppliedInput is not null)
             {
-                if(!SuppliedInput.EndsWith(".sln"))
+                if (!SuppliedInput.EndsWith(".sln"))
                 {
                     throw new InputException($"Given path is not a solution file: {SuppliedInput}");
                 }
@@ -26,7 +27,21 @@ namespace Stryker.Core.Options.Inputs
 
                 return fullPath;
             }
-            return null;
+            else
+            {
+                var solutionFiles = fileSystem.Directory.GetFiles(basePath, "*.*").Where(file => file.EndsWith("sln")).ToArray();
+                if (solutionFiles.Count() > 1)
+                {
+                    var sb = new StringBuilder();
+                    sb.AppendLine($"Expected exactly one .sln file, found more than one:");
+                    foreach (var file in solutionFiles)
+                    {
+                        sb.AppendLine(file);
+                    }
+                    throw new InputException(sb.ToString());
+                }
+                return solutionFiles.FirstOrDefault();
+            }
         }
     }
 }