+++ src/Stryker.Core/Stryker.Core/ToolHelpers/MsBuildHelper.cs
@@ -1,9 +1,9 @@
-using Microsoft.Extensions.Logging;
-using Stryker.Core.Logging;
-using Stryker.Core.Testing;
 using System.Collections.Generic;
 using System.IO;
 using System.IO.Abstractions;
+using Microsoft.Extensions.Logging;
+using Stryker.Core.Logging;
+using Stryker.Core.Testing;
 
 namespace Stryker.Core.ToolHelpers
 {
@@ -19,6 +19,7 @@ namespace Stryker.Core.ToolHelpers
             _logger = logger ?? ApplicationLogging.LoggerFactory.CreateLogger<MsBuildHelper>();
             fallbackLocations = new List<string>
             {
+                @"C:\Program Files (x86)\Microsoft Visual Studio\2017\Enterprise\MSBuild\15.0\Bin\MSBuild.exe",
                 @"C:\Windows\Microsoft.Net\Framework64\v4.0.30319\MSBuild.exe",
                 @"C:\Windows\Microsoft.Net\Framework64\v3.5\MSBuild.exe",
                 @"C:\Windows\Microsoft.Net\Framework64\v2.0.50727\MSBuild.exe",
@@ -30,10 +31,6 @@ namespace Stryker.Core.ToolHelpers
 
         public string GetMsBuildPath(IProcessExecutor _processExecutor)
         {
-            if (File.Exists(@"C:\Program Files (x86)\Microsoft Visual Studio\2017\Enterprise\MSBuild\15.0\Bin\MSBuild.exe"))
-            {
-                return @"C:\Program Files (x86)\Microsoft Visual Studio\2017\Enterprise\MSBuild\15.0\Bin\MSBuild.exe";
-            }
             // See if any MSBuild.exe can be found in visual studio installation folder
             foreach (string drive in Directory.GetLogicalDrives())
             {
@@ -58,7 +55,6 @@ namespace Stryker.Core.ToolHelpers
                     }
                 }
             }
-
             // Else, find in default locations
             _logger.LogDebug("Unable to find msbuild using vswhere, using fallback locations");