+++ src/Stryker.Core/CHANGELOG.md
@@ -1,3 +1,81 @@
+## [3.6.1](https://github.com/stryker-mutator/stryker-net/compare/stryker@3.6.0...stryker@3.6.1) (2023-02-17)
+
+
+### Features
+
+* **embedded resources:** Support embedded resources outside the root of the project ([#2414](https://github.com/stryker-mutator/stryker-net/issues/2414)) ([59fa56f](https://github.com/stryker-mutator/stryker-net/commit/59fa56f343aac9c837c21c5d4f122aee26c9318a))
+
+
+
+# [3.6.0](https://github.com/stryker-mutator/stryker-net/compare/stryker@3.5.1...stryker@3.6.0) (2023-02-10)
+
+
+### Bug Fixes
+
+* **embedded resources generation:** Use buildalyzer to find root namespace and embedded resources ([#2404](https://github.com/stryker-mutator/stryker-net/issues/2404)) ([a4969ff](https://github.com/stryker-mutator/stryker-net/commit/a4969ff60debcb6a94a059ba2d2aa84cb54029ef))
+
+
+### Features
+
+* **html report:** Update mutation testing elements to 1.7.14 ([#2405](https://github.com/stryker-mutator/stryker-net/issues/2405)) ([98ad817](https://github.com/stryker-mutator/stryker-net/commit/98ad8178e61c1fad54c208e2712063634a0f6961))
+
+
+
+## [3.5.1](https://github.com/stryker-mutator/stryker-net/compare/stryker@3.5.0...stryker@3.5.1) (2023-02-06)
+
+
+### Bug Fixes
+
+* **dashboard reporter:** fix dashboard publishing when the project name is computed ([#2397](https://github.com/stryker-mutator/stryker-net/issues/2397)) ([7395b2b](https://github.com/stryker-mutator/stryker-net/commit/7395b2bc4dfd9db962e1ea1e9d13bc7e6235c252)), closes [#1710](https://github.com/stryker-mutator/stryker-net/issues/1710)
+
+
+
+# [3.5.0](https://github.com/stryker-mutator/stryker-net/compare/stryker@3.4.0...stryker@3.5.0) (2023-02-06)
+
+
+### Bug Fixes
+
+* **no test projects:** null-reference exception when no test projects are found ([#2377](https://github.com/stryker-mutator/stryker-net/issues/2377)) ([419a8a7](https://github.com/stryker-mutator/stryker-net/commit/419a8a746190d2e4262781525b59aef7791fa56f))
+* **NullCoalescingExpressionMutator:** Don't generate certain mutants when right hand side is ThrowExpression ([#2387](https://github.com/stryker-mutator/stryker-net/issues/2387)) ([#2389](https://github.com/stryker-mutator/stryker-net/issues/2389)) ([e344382](https://github.com/stryker-mutator/stryker-net/commit/e34438259c26cf6b563c7a724b3f7bba14ddf768))
+
+
+### Features
+
+* **embedded resources:** Improve extraction of embedded resources ([#2383](https://github.com/stryker-mutator/stryker-net/issues/2383)) ([43c2918](https://github.com/stryker-mutator/stryker-net/commit/43c2918362a87be6f93b634df78c6ccb80bc6fbb))
+
+
+
+# [3.4.0](https://github.com/stryker-mutator/stryker-net/compare/stryker@3.3.0...stryker@3.4.0) (2022-12-23)
+
+
+### Bug Fixes
+
+* **nugetrestore:** use MsBuild option during nuget restore ([#2359](https://github.com/stryker-mutator/stryker-net/issues/2359)) ([892b69c](https://github.com/stryker-mutator/stryker-net/commit/892b69c1b0c9add9d4e35d8e65b21ac255510617))
+
+
+### Features
+
+* **pattern mutations:** Mutator for relational and logical patterns in is\switch expressions ([#2313](https://github.com/stryker-mutator/stryker-net/issues/2313)) ([70ded11](https://github.com/stryker-mutator/stryker-net/commit/70ded110c3dbf344386e012259c0d8a513ae9bbe))
+
+
+
+# [3.3.0](https://github.com/stryker-mutator/stryker-net/compare/stryker@3.2.0...stryker@3.3.0) (2022-12-09)
+
+
+### Bug Fixes
+
+* **GitDiffProvider:** remove stryker-config exclusion. ([#2338](https://github.com/stryker-mutator/stryker-net/issues/2338)) ([53286d3](https://github.com/stryker-mutator/stryker-net/commit/53286d34467d335175fd8660490525546cad1c14))
+* **Solution context run:** Allow relative path for solution file  ([#2333](https://github.com/stryker-mutator/stryker-net/issues/2333)) ([01fbf8a](https://github.com/stryker-mutator/stryker-net/commit/01fbf8a0ac56aa9bfc007eaeab6e9e3d954410cd))
+
+
+### Features
+
+* **mutators:** Add new methods to LinqMutator ([#2329](https://github.com/stryker-mutator/stryker-net/issues/2329)) ([bdec359](https://github.com/stryker-mutator/stryker-net/commit/bdec359a835e139b6632fb5b2da3a27c5a116ffc))
+* **mutators:** Add null-coalescing operator mutator ([#2284](https://github.com/stryker-mutator/stryker-net/issues/2284)) ([dc7c513](https://github.com/stryker-mutator/stryker-net/commit/dc7c513681559b31e4142e7fbbc1b9a7f271ff97))
+* **solution context run:** discover solution file ([#2340](https://github.com/stryker-mutator/stryker-net/issues/2340)) ([52bc645](https://github.com/stryker-mutator/stryker-net/commit/52bc645d9f16cc4f397045dcb1bf21b87da73283))
+
+
+
 # [3.2.0](https://github.com/stryker-mutator/stryker-net/compare/stryker@3.1.0...stryker@3.2.0) (2022-11-21)